integer lib mvp is implemented

+++ contracts/margined_vamm/src/state.rs
@@ -4,6 +4,8 @@ use serde::{Deserialize, Serialize};
 use cosmwasm_std::{Addr, StdResult, Storage, Timestamp, Uint128};
 use cosmwasm_storage::{bucket, bucket_read, singleton, singleton_read};
 
+use margined_common::integer::Integer;
+
 pub static KEY_CONFIG: &[u8] = b"config";
 pub static KEY_STATE: &[u8] = b"state";
 pub static KEY_RESERVE_SNAPSHOT: &[u8] = b"reserve_snapshot";
@@ -35,7 +37,7 @@ pub fn read_config(storage: &dyn Storage) -> StdResult<Config> {
 pub struct State {
     pub quote_asset_reserve: Uint128,
     pub base_asset_reserve: Uint128,
-    pub total_position_size: i128,
+    pub total_position_size: Integer,
     pub funding_rate: Uint128,
     pub next_funding_time: u64,
 }
